Fix CLI connections import
In connections_import, each connection was deserialized and stored into a
Connection model instance rather than a dictionary, so an erroneous call to the
dictionary methods .items() resulted in an AttributeError.  This fix accounts
for the fact that connections are loaded into SQLAlchemy models and accesses by
attribute rather than the dictionary method .items().