Adds script to automate renaming docker images from master to main
Since we are renaming `master` to `main` this script helps with
semi-automated syncing of `master` to `main` images in all the
Docker registries we are using:

* DockerHub - we keep `apache/airflow` image with distinct tags
  that determine type of the image, because in DockerHub we only
  have access to `apache/airflow` image

* GitHub Docker Registries: (depends on the type of registry) we have
  more flexibility:
  * In the old GitHub docker registry - docker.pkg.github.com -
    (current but already deprecated) we can use
    "apache/airflow/IMAGE:tag" i
  * in the new package registry (ghcr.io) - we can submitg anything
    under apache/airflow-* but then we link it to the
    project via docker image label.

The script helps to keep all the registries in-sync - copies
`master` to `main` so that we can run it to test the rename and
re-run it just before we switch the branches.

Also during implementation we found out that the documentation
about naming convetions had mistakes - this PR corrects them.