Set taskinstance state to the state reported by executor
When the executor reports a different taskinstance state to the one in DB, we capture the error
but fail to set the taskinstance state to the one reported by the executor with the effect that
the taskinstance ends up being in a successful state(in DB) even though it failed.

If for example, the first task in a dag is killed by deleting the task pod, the executor marks
the task as failed.
On getting to the scheduler, the scheduler sees the error but never changes
the state of the taskinstance to the one reported by the executor and at the end the taskinstance is marked
successful as well as the downstream taskinstances.

In the UI, the task appears to never have started and is marked correctly as failed while downstream tasks are skipped

This change ensures that if the task pod is killed while running. The correct state(failed) of the taskinstance is set as well as the correct state(upstream_failed) for the downstream taskinstances.